requirements.watch_system.txt
# Core dependencies
flask==2.3.3
flask-cors==4.0.0
requests==2.31.0
pymongo==4.5.0
python-dotenv==1.0.0
schedule==1.2.0

# Web scraping
beautifulsoup4==4.12.2
lxml==4.9.3
selenium==4.15.0

# Image processing and AI
Pillow==10.0.1
opencv-python-headless==4.8.1.78
scikit-learn==1.3.0
numpy==1.24.3

# Google Gemini AI
google-generativeai==0.3.2
google-cloud-aiplatform==1.36.0

# Vector database and embeddings
chromadb==0.4.15
transformers==4.35.0
torch==2.1.0
torchvision==0.16.0

# WhatsApp integration
twilio==8.9.1
pyngrok==7.0.0

# Database and caching
redis==5.0.0
sqlalchemy==2.0.23

# Utilities
python-dateutil==2.8.2
pytz==2023.3
tqdm==4.66.1
psutil==5.9.6

# Development and logging
pytest==7.4.3
black==23.9.1
flake8==6.1.0
python-multipart==0.0.6

# Additional ML libraries
sentence-transformers==2.2.2
faiss-cpu==1.7.4
langchain==0.0.325
openai==0.28.1

# Image similarity search
imagehash==4.3.1